WebsocketClient supports proxying
Added support for TLS and plain text proxying of Websocket traffic.

In the TLS case a CONNECT header is first sent establishing the proxied
traffic.

In the plain text case the modified URI in the initial upgrade request
header indicates to the proxy server that the traffic is to be proxied.

Use `NIOWebSocketFrameAggregator` to handle aggregating frame
fragments. This brings with it more protections e.g. against memory
exhaustion.

Accompanying config has been added to support this change.